Es un paso adelante?

+++ src/main/java/Modelo/Libro.java
@@ -1,5 +1,5 @@
 package Modelo;
 
-public record Libro(Long id,String title,String author,String genre) {
+public record Libro(int id,String title,String author,String genre) {
 
 }
